refactor: split mitm request handler tests
